modify motor steps

+++ HARDWARE/Cheminert/Cheminert_c52.h
@@ -9,7 +9,7 @@
 #include "uart_choose.h"
 #include "config.h"
 
-u8 cheminert_c52_c55_transfer(u8*tx_buf,u8 tx_size,u8*rx_buf, u8*rx_size,int timeout,bool wait_flag, int type_flag);
+u8 cheminert_c52_c55_transfer(u8*tx_buf,u8 tx_size,u8*rx_buf, int*rx_size,int timeout,bool wait_flag, int type_flag);
 
 typedef enum Command_Cheminert_type{ 
 		CHEMINERT_C52_CP=25,//actuator
